+++ src/lists.js
@@ -60,9 +60,9 @@
 Color, Point, WatcherMorph, StringMorph, SpriteMorph, ScrollFrameMorph,
 CellMorph, ArrowMorph, MenuMorph, snapEquals, Morph, isNil, localize, isString,
 MorphicPreferences, TableDialogMorph, SpriteBubbleMorph, SpeechBubbleMorph,
-TableFrameMorph, TableMorph, Variable, isSnapObject*/
+TableFrameMorph, TableMorph, Variable, isSnapObject, Costume*/
 
-modules.lists = '2019-April-27';
+modules.lists = '2019-June-04';
 
 var List;
 var ListWatcherMorph;
@@ -314,14 +314,20 @@ List.prototype.columnNames = function () {
     return [];
 };
 
-List.prototype.version = function (startRow, rows) {
+List.prototype.version = function (startRow, rows, startCol, cols) {
     var l = Math.min(startRow + rows, this.length()),
         v = this.lastChanged,
         r,
         i;
     for (i = startRow; i <= l; i += 1) {
         r = this.at(i);
-        v = Math.max(v, r.lastChanged ? r.lastChanged : 0);
+        if (r instanceof Costume) {
+            v = Math.max(v, r.version);
+        } else if (r instanceof List) {
+            v = Math.max(v, r.version(startCol, cols));
+        } else {
+            v = Math.max(v, r.lastChanged ? r.lastChanged : 0);
+        }
     }
     return v;
 };
@@ -677,12 +683,12 @@ ListWatcherMorph.prototype.init = function (list, parentCell) {
 ListWatcherMorph.prototype.update = function (anyway) {
     var i, idx, ceil, morphs, cell, cnts, label, button, max,
         starttime, maxtime = 1000;
-
     this.frame.contents.children.forEach(function (m) {
         if (m instanceof CellMorph) {
             if (m.contentsMorph instanceof ListWatcherMorph) {
                 m.contentsMorph.update();
-            } else if (isSnapObject(m.contents)) {
+            } else if (isSnapObject(m.contents) ||
+                    (m.contents instanceof Costume)) {
                 m.update();
             }
         }